Destroy method add to CMemFileMap

+++ ecal/core/src/io/ecal_memfile_db.cpp
@@ -31,10 +31,10 @@ namespace eCAL
 {
   CMemFileMap::~CMemFileMap()
   {
-    Cleanup();
+    Destroy();
   }
 
-  void CMemFileMap::Cleanup()
+  void CMemFileMap::Destroy()
   {
     // lock memory map access
     std::lock_guard<std::mutex> lock(m_memfile_map_mtx);
@@ -161,16 +161,19 @@ namespace eCAL
     {
       bool AddFile(const std::string& name_, const bool create_, const size_t len_, SMemFileInfo& mem_file_info_)
       {
+        if (!g_memfile_map()) return false;
         return g_memfile_map()->AddFile(name_, create_, len_, mem_file_info_);
       }
 
       bool RemoveFile(const std::string& name_, const bool remove_)
       {
+        if (!g_memfile_map()) return false;
         return g_memfile_map()->RemoveFile(name_, remove_);
       }
 
       bool CheckFileSize(const std::string& name_, const size_t len_, SMemFileInfo& mem_file_info_)
       {
+        if (!g_memfile_map()) return false;
         return g_memfile_map()->CheckFileSize(name_, len_, mem_file_info_);
       }
     }
